requirements.txt
flask
ollama
pdfplumber
python-docx
